refactor: PostBuild -> Copyer
